Enemies are able to collide with each other
Deleted some lines in enemy class that were moved to combo class.
Collision detection works. Rebecca and I were only able to get one of
the collided enemy to remove itself. It bugs out if we remove the second
one. Trying to have it replace the second one with new attributes and
image of the combined combo type.