Bug fixes!!!
Fixed a bug that caused the game to crash if there were greater than 5
enemies on screen and an enemy collided with a hero that would result in
a game loss. Also fixed a bug that combining enemies wasn't decrementing
the enemies left  in the level resulting in never winning.