Combination Complete!
Jasmine and I completed the combination system. The code may be buggy
because we added A LOT of new code so let us know if there is anything
that's wrong.
Also, we discovered that we can teleport up on the first lane and down
on the third. We haven't fixed it yet though.