Calculating score to currency and saving
Score is printed. Currency is the number at the bottom of the screen.
Saving and loading the currency is working. Need to make a clear
function to clear the currency (maybe). Also need to do some checking if
the item shop can access the currency variable and is able to deduct
the amount from currency.