Restructured levels into worlds
The levels in the level class are now in tables set by their world.
World is now passed everywhere. If you go to map from a victory or loss
screen it goes back to the world screen that you came from instead of
always going back to world 1. World 2 levels work now too.