Trash Item
Made a "compost" item to get rid of food on the screen once. Also added
comments in recipe book. I tested out the trash item and if two items
are too close, it gets rid of both. Will be fixing that. Took a while
since I had to understand the code.