Custom font set up + Set up for objective screen.
I have the "lobster two" font set up so it should adapt to all platforms
(I believe, I can only check it working for Windows obviously so we'd
have to double check this later). I believe you need to install the font
your personal computer to see the font running in the simulator so grab
the .ttf file from the resources folder and do so.

And I have the basic layout for the objective screen set up. Will finish
this up today.