Changed to Tapping for Attack(SAVE OLD CODE FIRST)
I saved the old auto attacking code into a folder in my computer. Please
do the same if you want to save the old code if we want to bring the
auto attacking back. I moved some functions out of the ingame file to
other files that are related so it's less cluttered. (ex: all the bullet
making functions are in heroes.lua and the score calculations in
score.lua) Also started on making the serving button. That doesn't work
yet since I was working on the tapping. I wasn't able to test the
serving if the food kept dying. I'm also pretty sure I tested everything
but if something breaks in relation to bullets/attacking, let me know
and I'll fix it right away! Also, there's no timer in the attacking
since I wasn't sure if we wanted that. I can add that in if we decide to
though.