Worked on getting the serve buttons more accurate
So now if there are no (breakfast) food in the (breakfast) lane, then
the button should be gone!

Known bugs in this build:
-if a (breakfast) food is in the (breakfast) lane and another lane at
the same time, the serve button flashes until you get both of them in
the (breakfast) lane. It's hard to press the button while it's flashing
unless you tap fast enough.
-if the level ends with a (dessert) food crashing into the (dessert)
chef then it says that the replayButton is trying to add a nil
EventListener.

Also removed the sparkles for now since it was having a nil crash again
if a (dessert) food crashes into the (dessert) chef and ends the level.