Level select in progress
Got the general layout but added in the title parameter in the level
class makes it break. I commented it out for now but it isn't able to
start the game when I hit "go". It only breaks at the line when I'm
telling the level select scene to require level.lua. Also added a new
background in the inBetween scene.

John and I tried adding the inBetween scene after selecting the level
but once it gets into the ingame, it breaks. :(