Lemon Assets + Stars and Score saving
Alexis worked really hard! ; v ; Also added the high score and saving
stars into world 2 and 3. No saving for item belt yet. That should be
the last one. Also needs a saved boolean for the recipe book and pantry
needs to be added so the player doesn't have to keep going through
tutorial.